requirements.txt
asgiref==3.3.4
Django==3.2.4
django-taggit==1.2.0
Markdown==3.3.4
psycopg2==2.8.6
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
